Consertado a Quantidade Utilizada da Matéria-Prima (#31)
* alterado o tipo da quantidade para Float

- não estava seguindo de acordo com o banco de dados
- tambem foi corrigido o nome de método isZeroOrNegative
- foi criado o IsZeroOrNegative(float) necessário
- corrigido os trechos de codigo com erro

* mudado a quantidade para float no JFrame

- foi alterado na criação do produto para quando o usuário for informar
  a quantidade da materia prima, permitir valores decimais

* atualizado os campos de visualizações para float

- corrigido a apresentação, estava convertendo para inteiro

* corrigido procedimentos e valores padrões DB

- foi arrumado a spCriarItensProduto para suportar o float da quantidade
- foram criadas materias primas, tipos e unidades de medidas adequadas

+++ MoonlightApp/src/main/java/com/moonlight/moonlightapp/models/ItemProdutoModel.java
@@ -7,23 +7,23 @@ package com.moonlight.moonlightapp.models;
  */
 
 public class ItemProdutoModel extends BaseModel {
-    private int quantidade;
+    private Float quantidade;
     private ProdutoModel produto;
     private MateriaPrimaModel materiaPrima;
-    private double subTotal;
+    private Double subTotal;
 
-    public ItemProdutoModel(int quantidade, ProdutoModel produto, double subTotal, MateriaPrimaModel materiaPrima) {
+    public ItemProdutoModel(Float quantidade, ProdutoModel produto, Double subTotal, MateriaPrimaModel materiaPrima) {
         this.quantidade = quantidade;
         this.produto = produto;
         this.subTotal = subTotal;
         this.materiaPrima = materiaPrima;
     }
 
-    public int getQuantidade() {
+    public Float getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(int quantidade) {
+    public void setQuantidade(Float quantidade) {
         this.quantidade = quantidade;
     }
 
@@ -43,11 +43,11 @@ public class ItemProdutoModel extends BaseModel {
         this.materiaPrima = materiaPrima;
     }
 
-    public double getSubTotal() {
+    public Double getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(double subTotal) {
+    public void setSubTotal(Double subTotal) {
         this.subTotal = subTotal;
     }
 }
